separate flock implementations into separate modules
